// WebCore/apps/ds-source/src/app/admin/ar/shared/ar-report-type.ts
export enum ArReportType {
    BillingDetail = 1,
    BaseRevenueList = 2,
    DepositDetail = 3,
    DepositSummary = 4,
    AgingReport = 5,
    SalesByDate = 6,
    ClientSales = 7,
    ManualBillingInvoice = 8,
    CreditListing = 9,
    GainsAndLosses = 10,
    BillingInvoice = 11
}